Update users of the File class
The previous patch extended the File (and related) classes to support 64-bit
filesizes. This patch changes the users of the File class. Often this change is
simply not using 32-bit variable types to store the filesize. In a few cases
(e.g. ROM or XML files) I still added a check for a maximum filesize because
adapting the whole underlying code was a lot of work and/or not worth the
effort.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13086 7512d703-071c-0410-b5c2-d2bdd895b7d1

+++ src/cassette/CasImage.hh
@@ -33,7 +33,7 @@ private:
 	void writeHeader(int s);
 	void writeSilence(int s);
 	void writeByte(byte b);
-	bool writeData(const byte* buf, unsigned size, unsigned& pos);
+	bool writeData(const byte* buf, size_t size, size_t& pos);
 	void convert(const Filename& filename, FilePool& filePool, CliComm& cliComm);
 
 	std::vector<signed char> output;
